fix test cases for connection errors
